requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
itsdangerous>=2.2,<3.0
pillow>=10.0,<13.0
